requirements.txt
numpy==1.24.4 
Pillow==10.3.0 
Requests==2.31.0 
torch==2.3.0 
torchvision==0.18.0 
transformers==4.40.2
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118
accelerate
bitsandbytes
git+https://github.com/huggingface/transformers
